multi-agent-autotest/backend/requirements.txt
# API
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9  # upload de arquivos

# Agentes
langgraph==0.2.28
langchain-openai==0.2.3

# Utilitários
pydantic==2.9.2
python-dotenv==1.0.1
jinja2==3.1.4